Update site to work in "archive mode"
When built or run with the environment variable `ARCHIVE_MODE` set to
"true", the site will mount itself under the `/2025/heist/` (with the
expected hostname of `puzzles.mit.edu`).

The changes to support this look somewhat extensive, but mostly consist
of prefixing all URLs with `import.meta.env.BASE_URL`. These
replacements should be exhaustive (I'm pretty sure I found any string
that started with `/`), although it's certainly possible I missed
something.

Since puzzles.mit.edu is a shared space, I also added a prefix to
localStorage keys (when `ARCHIVE_MODE` is enabled) to avoid collisions
with other archived hunts. The reset function is also adjusted in
`ARCHIVE_MODE` to avoid clearing keys that do not have the prefix.

All of these changes should be noops when `ARCHIVE_MODE` is not set,
allowing the site to continue running normally on the
mitmysteryheist.com domain.